Draw It or Lose It Software Design Template
The Gaming Room currently has a game application, Draw It or Lose It, that was created for Android systems. They would like to expand it to other operating platforms. To do so, I have been asked to create a simulated environment that demonstrates prototypes to present multiple platform options to my client. I have created a Software Design Template to satisfy all requirements by The Gaming Room.